Move TypeTest dataProviders inside of tests
Since the dataProvider was the thing invoking the Type static factory
methods, those calls didn't count towards code coverage. Since
infection uses code coverage to determine which tests to re-run when it
makes a mutation, that meant that it didn't run the important tests when
checking if the mutation broke any tests.

Upped the min msi values again to match what it was before