Throw JsonFormatException when field should be an array but is an object
There was a bug that caused it to throw TypeError instead

Depending on your definition, this may be a breaking change. In the case
that you noticed this bug and caught TypeError but not
JsonFormatException, this would break your code. That seems incredibly
unlikely though, it seems far more likely that the harm of pushing this
change to a major version change is greater